Move exception handler to Infrastructure namespace
